hellgames/requirements.txt
asgiref==3.2.10
boto3==1.14.44
botocore==1.17.44
dj-database-url==0.5.0
Django==3.0.6
django-filter==2.3.0
django-generic-positions==0.2.3
django-hvad==1.8.0
django-libs==2.0.3
django-money==1.1
django-phonenumber-field==5.0.0
django-storages==1.9.1
django-user-media==1.2.5
docutils==0.15.2
easy-thumbnails==2.7
gunicorn==20.0.4
jmespath==0.10.0
phonenumbers==8.12.11
Pillow==7.2.0
psycopg2==2.8.5
py-moneyed==0.8.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
